added test cases of addnic
